requirements.txt
aiogram==3.2.0
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
apscheduler==3.10.4
openai>=1.50.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.24.1
httpx-socks==0.7.7
markdown==3.5.1